backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
requests==2.32.3
redis==5.0.8
cryptography==43.0.1
alembic==1.13.2
slowapi==0.1.9
httpx==0.27.2
authlib==1.3.2
python-multipart==0.0.9
email-validator==2.2.0
